LayoutTests/http/wpt/beacon/beacon-readablestream.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    test(function() {
        assert_throws_js(TypeError, () => { navigator.sendBeacon("/", new ReadableStream()) });
    }, "Beacon with a ReadableStream body should fail.");
</script>
